+++ src/ts/static/linksIcons.ts
@@ -1,0 +1,20 @@
+const linksIcons = [
+    {
+        linkTitle: 'The PHP Foundation',
+        iconName: 'phpfoundation.png',
+    },
+    {
+        linkTitle: 'Upcoming conferences',
+        iconName: 'talk.png',
+    },
+    {
+        linkTitle: 'User Group Events',
+        iconName: 'events.png',
+    },
+    {
+        linkTitle: 'Special Thanks',
+        iconName: 'thanks.png',
+    },
+]
+
+export default linksIcons