+++ src/ts/static/socialLinks.ts
@@ -1,0 +1,11 @@
+import type { HomeLink } from '@/types'
+
+const recommendedLinks: HomeLink[] = [
+    {
+        href: 'https://x.com/official_php',
+        iconName: 'x.svg',
+        title: 'PHP Official',
+    },
+]
+
+export default recommendedLinks