+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## v2.8.1 (2024-04-29)
+## v2.8.1 (2024-04-30)
 - Change the 'dollar' icon for the 'Donate' button on the home page from png to svg
 - Added a hover effect for each recommended video on the home page sidebar
 - Changed logic with loading recommended videos on the home page